// Frontend/src/app/models/release.ts
export class Release {
    idrelease: number = 0;
    idstatus: number = 0;
    idaplicatie: number = 0;
    idmediu: number = 0;
    idutilizator: number = 0;
    iddurata: number = 0;
    imbunatatiri: number = 0;
    bugs: number = 0;
    esteurgenta: number = 0;
    contentrelease: string = '';
    comentarii: string = '';
    brvpath: string = '';
    testpath: string = '';
    datacreare: Date = new Date();
    datamodificare: Date = new Date();
    relstatus: number = 0;
}